feature: split main.py into 2 files game.py and utils.py + refactoring Makefile targets
