+++ pyproject.toml
@@ -20,6 +20,9 @@ shifumi = "main:main_cli"
 # requires = ["hatchling"]
 # build-backend = "hatchling.build"
 
+[tool.uv]
+package = true
+
 [tool.setuptools]
 include-package-data = false
