+++ src/utils.py
@@ -1,0 +1,6 @@
+import pygame
+
+
+# simple wrapper to keep the screen resizeable
+def init_screen(width, height):
+    return pygame.display.set_mode((width, height), pygame.RESIZABLE)
